+++ feature_listen_later_ui/src/main/java/com/allsoftdroid/audiobook/feature_listen_later_ui/di/FeatureListenLaterModule.kt
@@ -1,7 +1,7 @@
 package com.allsoftdroid.audiobook.feature_listen_later_ui.di
 
 import androidx.annotation.VisibleForTesting
-import com.allsoftdroid.audiobook.feature_listen_later_ui.ListenLaterViewModel
+import com.allsoftdroid.audiobook.feature_listen_later_ui.presentation.ListenLaterViewModel
 import com.allsoftdroid.database.common.AudioBookDatabase
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
