+++ feature_listen_later_ui/src/main/java/com/allsoftdroid/audiobook/feature_listen_later_ui/presentation/ListenLaterFragment.kt
@@ -1,10 +1,11 @@
-package com.allsoftdroid.audiobook.feature_listen_later_ui
+package com.allsoftdroid.audiobook.feature_listen_later_ui.presentation
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.activity.OnBackPressedCallback
+import com.allsoftdroid.audiobook.feature_listen_later_ui.R
 import com.allsoftdroid.audiobook.feature_listen_later_ui.databinding.FragmentListenLaterLayoutBinding
 import com.allsoftdroid.audiobook.feature_listen_later_ui.di.FeatureListenLaterModule
 import com.allsoftdroid.common.base.fragment.BaseUIFragment
@@ -12,7 +13,7 @@ import org.koin.android.ext.android.inject
 
 class ListenLaterFragment : BaseUIFragment() {
 
-    val listenLaterViewModel:ListenLaterViewModel by inject()
+    val listenLaterViewModel: ListenLaterViewModel by inject()
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -20,7 +21,8 @@ class ListenLaterFragment : BaseUIFragment() {
         savedInstanceState: Bundle?
     ): View? {
 
-        val dataBinding:FragmentListenLaterLayoutBinding = inflateLayout(inflater,R.layout.fragment_listen_later_layout,container,false)
+        val dataBinding:FragmentListenLaterLayoutBinding = inflateLayout(inflater,
+            R.layout.fragment_listen_later_layout,container,false)
         FeatureListenLaterModule.unLoadModules()
 
         return dataBinding.root
