+++ feature_listen_later_ui/src/main/res/layout/fragment_listen_later_layout.xml
@@ -4,7 +4,7 @@
 
     <data>
         <variable name="viewModel"
-            type="com.allsoftdroid.audiobook.feature_listen_later_ui.ListenLaterViewModel"/>
+            type="com.allsoftdroid.audiobook.feature_listen_later_ui.presentation.ListenLaterViewModel"/>
     </data>
 
     <LinearLayout
